[fix:gradle.build][add:CreateSwimmerTest.java]
